+++ src/main/java/shareddtos/billmodule/bill/ShowBillDto.java
@@ -1,6 +1,5 @@
 package shareddtos.billmodule.bill;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -38,11 +37,11 @@ public class ShowBillDto {
     private List<BillItemDTO> products;
 
     // Totals
-    private BigDecimal nonTaxedSales;
-    private BigDecimal exemptSales;
-    private BigDecimal taxedSales;
-    private BigDecimal iva; 
-    private BigDecimal perceivedIva;
-    private BigDecimal withheldIva;
-    private BigDecimal totalWithIva;
+    private Double nonTaxedSales;
+    private Double exemptSales;
+    private Double taxedSales;
+    private Double iva; 
+    private Double perceivedIva;
+    private Double withheldIva;
+    private Double totalWithIva;
 }
